New JS File (week 3 hw)
JS file: prompted users to search for desired city's weather. file included info properties of 5 specific cities. if searched city was not in code or is user had not responded then page would alert accordingly.

Modified the CSS to center the card group and main section.

Added the new JS file to HTML.
<script src=app.js></script>